requirements.txt
Flask==1.1.2
hdbscan==0.8.26
numpy==1.18.2
pandas==1.0.3
PyYAML==5.3.1
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
spacy==2.2.4
umap-learn==0.4.1 
python-dotenv  
